requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-static==0.0.6
Django==3.1.1
djangorestframework==3.11.1
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.10
